Moved CheckerHelperFunctions one folder up (as it was misplaced in exceptions folder) Removed exceptions folder for checkers with old MismatchException.java as it is no longer needed Moved executeSetter functionality to CheckerHelperFunctions.java
