--- src/main/java/de/a9d3/testing/checker/exception/MismatchException.java
@@ -1,7 +1,0 @@
-package de.a9d3.testing.checker.exception;
-
-public class MismatchException extends RuntimeException {
-    public MismatchException(String message) {
-        super(message);
-    }
-}
